python/requirements.txt
opencv-python-headless
numpy
tensorflow-cpu